Split BagManifest into two separate types, and the parser too
Within the context of a BagIt bag, there is a difference between
the payload manifest and the tag manifest, even though they both have
the same format.  Enforce that difference in the type system, so we
can't mix the two up!
