+++ common/src/main/scala/uk/ac/wellcome/platform/archive/common/bagit/models/Bag.scala
@@ -2,7 +2,7 @@ package uk.ac.wellcome.platform.archive.common.bagit.models
 
 case class Bag(
   info: BagInfo,
-  manifest: BagManifest,
-  tagManifest: BagManifest,
+  manifest: PayloadManifest,
+  tagManifest: TagManifest,
   fetch: Option[BagFetch] = None
 )
